+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Added aliases for most commands (eg. `deploy,d`, `restart,rr`, `run,r`, `stop,s`).
+- Added two new options for logs `--all` and `--follow`.
+
 
 ## [0.3.1-alpha] - 2020-01-11
 ### Added